Input: Moved input to new file
