+++ nightmare.vcxproj
@@ -160,6 +160,7 @@ copy /y $(SolutionDir)..\sdl2_image\lib\x86\*.dll $(OutputPath)</Command>
     <ClCompile Include="src\blob.cc" />
     <ClCompile Include="src\char.cc" />
     <ClCompile Include="src\chunk.cc" />
+    <ClCompile Include="src\input.cc" />
     <ClCompile Include="src\mesh.cc" />
     <ClCompile Include="src\mesher.cc" />
     <ClCompile Include="src\physics.cc" />
@@ -175,6 +176,7 @@ copy /y $(SolutionDir)..\sdl2_image\lib\x86\*.dll $(OutputPath)</Command>
     <ClInclude Include="src\common.h" />
     <ClInclude Include="src\fixed_cube.h" />
     <ClInclude Include="src\fixed_grid.h" />
+    <ClInclude Include="src\input.h" />
     <ClInclude Include="src\mesh.h" />
     <ClInclude Include="src\physics.h" />
     <ClInclude Include="src\player.h" />
